pyproject.toml
[project]
name = "ucp-conformance-tests"
version = "0.1.0"
description = "Test the conformance of a merchant server implementation with the UCP spec"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "UCP Team" }
]
dependencies = [
    "absl-py>=1.0.0",
    "httpx>=0.26.0",
    "pydantic>=2.12.0",
    "ucp-sdk",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.38.0",
    "ruff>=0.14.11",
]

[dependency-groups]
dev = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv.sources]
# The relative path is stored here
ucp-sdk = { path = "../python-sdk/", editable = true }

[tool.ruff]
line-length = 80
indent-width = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

[tool.ruff.lint]
select = ["E", "F", "W", "B", "C4", "SIM", "N", "UP", "D", "PTH", "T20"]
ignore = ["D203", "D213"]

[tool.ruff.lint.isort]
combine-as-imports = true
force-sort-within-sections = true
case-sensitive = true
